Refactor tests for events and sessions.
